+++ src/main/java/com/example/ecommerce_online/model/entity/Price.java
@@ -7,6 +7,7 @@ import javax.persistence.Entity;
 import javax.persistence.Id;
 import javax.persistence.Table;
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 @Setter
 @Getter
@@ -19,5 +20,5 @@ import java.io.Serializable;
 public class Price implements Serializable {
     @Id
     Long id;
-    Long price;
+    BigDecimal price;
 }
